Mini fix to last commit
