+++ fe1-web/components/eventList/events/index.ts
@@ -1,0 +1,3 @@
+import Event from './Event';
+
+export default Event;
